refactor: Move eof and value to the item module
